Performance improvements for displaying charts
Splits the observables into partitions of 10 at a time, ordered by count
to show the highest first. View models with a large number of
observables would be very difficult to show on one graph. The next step
may be to hide the graphs that aren't being currently looked at and
provide a mechanism to cycle between them.